requirements.txt
absl-py==1.4.0
antlr4-python3-runtime==4.9.3
asttokens==2.2.1
backcall==0.2.0
black==23.3.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
-e git+https://github.com/facebookresearch/detectron2.git@d779ea63faa54fe42b9b4c280365eaafccb280d6#egg=detectron2
executing==1.2.0
filelock==3.10.7
fonttools==4.39.3
fvcore==0.1.5.post20221221
google-auth==2.17.1
google-auth-oauthlib==1.0.0
grpcio==1.53.0
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.1.0
importlib-resources==5.12.0
iopath==0.1.9
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.0
numpy==1.24.2
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathspec==0.11.1
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
portalocker==2.7.0
prompt-toolkit==3.0.38
protobuf==4.21.2
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
Pygments==2.14.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
tabulate==0.9.0
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
thop==0.1.1.post2209072238
tomli==2.0.1
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.2.3
yacs==0.1.8
zipp==3.15.0
